Refactor SelectColumnOrder and SelectColumn components
